Removing PAssertExtension
I've been thinking about this - i've come to think that frameworks that add extension methods to object are a bit rude. Especially when using resharper.
My recommendation is that this class becomes a code snippet that anyone can add to their own test library if they want the syntactic sugar